+++ internal/test/e2e/otelglobal/main.go
@@ -1,0 +1,55 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package main
+
+import (
+	"context"
+	"time"
+
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+)
+
+var tracer = otel.Tracer("trace-example")
+
+func innerFunction(ctx context.Context) {
+	_, span := tracer.Start(ctx, "child")
+	defer span.End()
+
+	span.SetAttributes(attribute.String("inner.key", "inner.value"))
+}
+
+func createMainSpan(ctx context.Context) {
+	ctx, span := tracer.Start(ctx, "parent")
+	defer span.End()
+
+	innerFunction(ctx)
+
+	intAttr := attribute.Int("int_key", 42)
+	strAttr := attribute.String("string_key", "forty-two")
+	boolAttr := attribute.Bool("bool_key", true)
+	floatAttr := attribute.Float64("float_key", 42.3)
+	span.SetAttributes(intAttr, strAttr, boolAttr, floatAttr)
+}
+
+func main() {
+	// give time for auto-instrumentation to start up
+	time.Sleep(5 * time.Second)
+
+	createMainSpan(context.Background())
+
+	// give time for auto-instrumentation to report signal
+	time.Sleep(5 * time.Second)
+}